Move get timestamp function to SystemInfo singleton
