+++ src/cpp/fastdds/log/Log.cpp
@@ -23,6 +23,7 @@
 #include <fastdds/dds/log/StdoutConsumer.hpp>
 #include <fastdds/dds/log/StdoutErrConsumer.hpp>
 #include <fastdds/dds/log/Colors.hpp>
+#include <utils/SystemInfo.hpp>
 #include <iostream>
 
 using namespace std;
@@ -275,8 +276,7 @@ void Log::QueueLog(
         }
     }
 
-    std::string timestamp;
-    get_timestamp(timestamp);
+    std::string timestamp = SystemInfo::get_timestamp();
     resources_.logs.Push(Log::Entry{message, context, kind, timestamp});
     {
         std::unique_lock<std::mutex> guard(resources_.cv_mutex);
@@ -316,32 +316,6 @@ void Log::SetErrorStringFilter(
 {
     std::unique_lock<std::mutex> configGuard(resources_.config_mutex);
     resources_.error_string_filter.reset(new std::regex(filter));
-}
-
-void Log::get_timestamp(
-        std::string& timestamp)
-{
-    std::stringstream stream;
-    auto now = std::chrono::system_clock::now();
-    std::time_t now_c = std::chrono::system_clock::to_time_t(now);
-    std::chrono::system_clock::duration tp = now.time_since_epoch();
-    tp -= std::chrono::duration_cast<std::chrono::seconds>(tp);
-    auto ms = static_cast<unsigned>(tp / std::chrono::milliseconds(1));
-
-#if defined(_WIN32)
-    struct tm timeinfo;
-    localtime_s(&timeinfo, &now_c);
-    //#elif defined(__clang__) && !defined(std::put_time) // TODO arm64 doesn't seem to support std::put_time
-    //    (void)now_c;
-    //    (void)ms;
-#elif defined(__unix__)
-    std::tm timeinfo;
-    localtime_r(&now_c, &timeinfo);
-#else
-    std::tm timeinfo = *localtime(&now_c);
-#endif // if defined(_WIN32)
-    stream << std::put_time(&timeinfo, "%F %T") << "." << std::setw(3) << std::setfill('0') << ms << " ";
-    timestamp = stream.str();
 }
 
 void LogConsumer::print_timestamp(
